+++ src/components/Button/Button.tsx
@@ -37,6 +37,7 @@ const Button = (props: ButtonProps) => {
         img,
         onClick: onClickOrigin,
         text,
+        width,
         ...rest
     } = props;
     const defaultImgPosition = 'left';
@@ -53,14 +54,16 @@ const Button = (props: ButtonProps) => {
         [handleMetrika, metrikaGoals, pixelEvents, handleAnalytics, analyticsEvents, onClickOrigin],
     );
 
+    const buttonModifiers = {size, theme, width};
+
     const buttonImg =
         img instanceof Object
             ? {url: img.url, position: img.position || defaultImgPosition, alt: img.alt}
             : {url: img, position: defaultImgPosition};
 
     const buttonClass = img
-        ? b({position: buttonImg.position, size, theme}, className)
-        : b({size, theme}, className);
+        ? b({position: buttonImg.position, ...buttonModifiers}, className)
+        : b({...buttonModifiers}, className);
 
     const buttonProps = {
         ...rest,
@@ -91,6 +94,7 @@ const Button = (props: ButtonProps) => {
             view={toCommonView(buttonTheme as OldButtonTheme)}
             size={toCommonSize(size as OldButtonSize)}
             href={url ? setUrlTld(url, tld) : undefined}
+            width={width}
             {...buttonProps}
         >
             {icon && buttonImg.position === 'left' ? icon : null}