+++ src/containers/PageConstructor/components/ConstructorBlocks/ConstructorBlocks.tsx
@@ -2,6 +2,7 @@ import React, {Fragment, ReactElement, useContext} from 'react';
 
 import _ from 'lodash';
 
+import {BlockIdContext} from '../../../../context/blockIdContext';
 import {InnerContext} from '../../../../context/innerContext';
 import {Block, ConstructorItem as ConstructorItemType} from '../../../../models';
 import {getBlockKey} from '../../../../utils';
@@ -57,18 +58,20 @@ export const ConstructorBlocks: React.FC<ConstructorBlocksProps> = ({items}) => 
             }
 
             itemElement = (
-                <ConstructorItem data={item} key={blockId} blockKey={blockId}>
+                <ConstructorItem data={item} blockKey={blockId}>
                     {children}
                 </ConstructorItem>
             );
         }
 
-        return blockTypes.includes(item.type) ? (
-            <ConstructorBlock data={item as Block} key={blockId}>
-                {itemElement}
-            </ConstructorBlock>
-        ) : (
-            itemElement
+        return (
+            <BlockIdContext.Provider value={blockId} key={blockId}>
+                {blockTypes.includes(item.type) ? (
+                    <ConstructorBlock data={item as Block}>{itemElement}</ConstructorBlock>
+                ) : (
+                    itemElement
+                )}
+            </BlockIdContext.Provider>
         );
     };