+++ src/containers/PageConstructor/components/ConstructorItem/ConstructorItem.tsx
@@ -1,7 +1,7 @@
 import React, {useContext} from 'react';
 
-import {BlockIdContext} from '../../../../context/blockIdContext';
 import {InnerContext} from '../../../../context/innerContext';
+import {BlockDecoration} from '../../../../customization/BlockDecoration';
 import {ConstructorItem as ConstructorItemType, WithChildren} from '../../../../models';
 
 export interface ConstructorItemProps {
@@ -9,7 +9,7 @@ export interface ConstructorItemProps {
     blockKey: string;
 }
 
-export const ConstructorItem = ({data, blockKey, children}: WithChildren<ConstructorItemProps>) => {
+export const ConstructorItem = ({data, children}: WithChildren<ConstructorItemProps>) => {
     const {itemMap} = useContext(InnerContext);
     const {type, ...rest} = data;
 
@@ -17,16 +17,14 @@ export const ConstructorItem = ({data, blockKey, children}: WithChildren<Constru
         React.ComponentProps<typeof itemMap[typeof type]>
     >;
 
-    return (
-        <BlockIdContext.Provider value={blockKey}>
-            <Component {...rest}>{children}</Component>
-        </BlockIdContext.Provider>
-    );
+    return <Component {...rest}>{children}</Component>;
 };
 
 export const ConstructorHeader = ({
     data,
     blockKey,
 }: Pick<ConstructorItemProps, 'data' | 'blockKey'>) => (
-    <ConstructorItem data={data} key={data.type} blockKey={blockKey} />
+    <BlockDecoration id={data.type}>
+        <ConstructorItem data={data} key={data.type} blockKey={blockKey} />
+    </BlockDecoration>
 );