+++ src/utils/blocks.ts
@@ -1,4 +1,4 @@
-import {CustomConfig, PCShareSocialNetwork, TextSize} from '../models';
+import {Block, CustomConfig, PCShareSocialNetwork, TextSize} from '../models';
 import {ConstructorBlock} from '../models/constructor';
 
 export function getHeaderTag(size: TextSize) {
@@ -54,6 +54,14 @@ const getShareUrlWithParams = (url: string, params: Record<string, string | unde
     return result.toString();
 };
 
+export const getOrderedBlocks = (blocks: Block[], headerBlockTypes: string[] = []) => {
+    return blocks.filter(({type}) => !headerBlockTypes.includes(type));
+};
+
+export const getHeaderBlock = (blocks: Block[], headerBlockTypes: string[] = []) => {
+    return blocks.find(({type}) => headerBlockTypes.includes(type));
+};
+
 export const getShareLink = (
     url: string,
     type: PCShareSocialNetwork,
@@ -88,3 +96,6 @@ export const getShareLink = (
             return undefined;
     }
 };
+
+export const getBlockIndexFromId = (blockId?: string) =>
+    Number(blockId?.split('-')?.slice(-1).pop());